TST: Consolidate numpy/matrixlib/tests.
The test files test_multiarray.py, test_numeric.py, and test_regression.py
contained very few tests, so merge them with test_defmatrix.py. As the
names of the three merged files are duplicated elsewhere, this also
helps to achieve the goal of every test file having a unique name,
simplifying the transition to pytest.
